Refactor: Delegate UI logic and "Remember-Bloat" to sub-components #193
